+++ .idea/modules/swith.main.iml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$/../../Swith-Server/build/generated/querydsl/generated">
+      <sourceFolder url="file://$MODULE_DIR$/../../Swith-Server/build/generated/querydsl/generated" isTestSource="false" generated="true" />
+    </content>
+  </component>
+</module>
