+++ Swith-Server/src/main/java/com/swith/api/application/dto/PatchApplicationStatusReq.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.request;
+package com.swith.api.application.dto;
 
 
 import lombok.Data;
