+++ Swith-Server/src/main/java/com/swith/api/groupinfo/dto/PatchEndGroupReq.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.request;
+package com.swith.api.groupinfo.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
