+++ Swith-Server/src/main/java/com/swith/api/groupinfo/dto/PostGroupInfoRes.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto;
+package com.swith.api.groupinfo.dto;
 
 
 import io.swagger.annotations.ApiModelProperty;
