+++ Swith-Server/src/main/java/com/swith/domain/attendance/service/AttendanceService.java
@@ -4,6 +4,7 @@ import com.swith.api.dto.response.*;
 import com.swith.domain.attendance.entity.Attendance;
 import com.swith.domain.attendance.repository.AttendanceRepository;
 import com.swith.domain.groupinfo.entity.GroupInfo;
+import com.swith.domain.groupinfo.repository.GroupInfoRepository;
 import com.swith.domain.register.entity.Register;
 import com.swith.domain.register.repository.RegisterRepository;
 import com.swith.domain.session.entity.Session;
@@ -13,7 +14,6 @@ import com.swith.domain.user.repository.UserRepository;
 import com.swith.global.error.exception.BaseException;
 import com.swith.api.common.constant.BaseResponseStatus;
 import com.swith.api.dto.request.PatchAttendanceReq;
-import com.swith.src.repository.*;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
