+++ Swith-Server/src/main/java/com/swith/domain/groupinfo/repository/GroupInfoRepository.java
@@ -1,4 +1,4 @@
-package com.swith.src.repository;
+package com.swith.domain.groupinfo.repository;
 
 
 import com.swith.domain.groupinfo.entity.GroupInfo;
