+++ Swith-Server/src/main/java/com/swith/domain/groupinfo/repository/GroupInfoRepositoryImpl.java
@@ -1,6 +1,6 @@
-package com.swith.src.repository;
+package com.swith.domain.groupinfo.repository;
 
-import com.swith.api.dto.request.GetGroupInfoSearchReq;
+import com.swith.api.groupinfo.dto.GetGroupInfoSearchReq;
 import com.swith.api.dto.response.GetGroupInfoSearchRes;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.ExpressionUtils;
