+++ Swith-Server/src/main/java/com/swith/domain/groupinfo/service/GroupInfoService.java
@@ -1,10 +1,12 @@
-package com.swith.src.service;
+package com.swith.domain.groupinfo.service;
 
 import com.swith.domain.announcement.entity.Announcement;
 import com.swith.domain.announcement.repository.AnnouncementRepository;
+import com.swith.domain.application.repository.ApplicationRepository;
 import com.swith.domain.attendance.entity.Attendance;
 import com.swith.domain.attendance.repository.AttendanceRepository;
 import com.swith.domain.groupinfo.entity.GroupInfo;
+import com.swith.domain.groupinfo.repository.GroupInfoRepository;
 import com.swith.domain.interest.entity.Interest;
 import com.swith.domain.interest.repository.InterestRepository;
 import com.swith.domain.register.entity.Register;
@@ -15,16 +17,15 @@ import com.swith.domain.user.repository.UserRepository;
 import com.swith.domain.session.repository.SessionRepository;
 import com.swith.global.error.exception.BaseException;
 import com.swith.api.common.constant.BaseResponseStatus;
-import com.swith.api.dto.request.PatchGroupInfoReq;
+import com.swith.api.groupinfo.dto.PatchGroupInfoReq;
 import com.swith.api.dto.response.GetEachGroupInfoRes;
 import com.swith.api.dto.response.GetHomeGroupInfoRes;
-import com.swith.api.dto.PostGroupInfoReq;
-import com.swith.api.dto.PostGroupInfoRes;
-import com.swith.api.dto.request.GetGroupInfoSearchReq;
+import com.swith.api.groupinfo.dto.PostGroupInfoReq;
+import com.swith.api.groupinfo.dto.PostGroupInfoRes;
+import com.swith.api.groupinfo.dto.GetGroupInfoSearchReq;
 import com.swith.api.dto.response.GetGroupInfoSearchRes;
 import com.swith.external.firebase.FirebaseCloudMessageService;
 import com.querydsl.jpa.impl.JPAQuery;
-import com.swith.src.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
