+++ Swith-Server/src/main/java/com/swith/domain/session/service/SessionService.java
@@ -5,6 +5,7 @@ import com.swith.domain.announcement.repository.AnnouncementRepository;
 import com.swith.domain.attendance.entity.Attendance;
 import com.swith.domain.attendance.repository.AttendanceRepository;
 import com.swith.domain.groupinfo.entity.GroupInfo;
+import com.swith.domain.groupinfo.repository.GroupInfoRepository;
 import com.swith.domain.memo.entity.Memo;
 import com.swith.domain.memo.repository.MemoRepository;
 import com.swith.domain.register.repository.RegisterRepository;
@@ -19,8 +20,7 @@ import com.swith.api.dto.response.GetSessionRes;
 import com.swith.api.dto.request.PostSessionReq;
 import com.swith.api.dto.response.GetSessionTabRes;
 import com.swith.api.dto.response.SessionAttendanceInfo;
-import com.swith.src.repository.*;
-import com.swith.src.service.GroupInfoService;
+import com.swith.domain.groupinfo.service.GroupInfoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
